Add files via upload
MyClass2.C is v2.0 of a MakeClass run on B5.root. This is under development and currently outputs the electromagnetic calorimeter energy deposition as a test histogram, as well as the reconstructed xz-plane trajectories for DC1 and DC2. This will be modified to reconstruct momenta. MyClass.h is the header file.